requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cot_reports==0.1.3
debugpy==1.8.1
decorator==5.1.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.0
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.19.0
